% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_data}
\alias{batch_data}
\title{conditionning input data into tensors according to tft variable roles}
\usage{
batch_data(recipe, df, total_time_steps = 12, device)
}
\arguments{
\item{recipe}{a recipe affecting tft roles to df}

\item{df}{a data frame}

\item{total_time_steps}{time_step value (default 48)}

\item{device}{the device to use for training. \code{cpu} or \code{cuda}. The default (\code{auto})
uses \code{cuda} if it's available, otherwise uses \code{cpu}.}
}
\description{
conditionning input data into tensors according to tft variable roles
}
