% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{covariates_spec}
\alias{covariates_spec}
\title{Create a specification of covariates types}
\usage{
covariates_spec(index, keys, static = NULL, known = NULL, unknown = NULL)
}
\arguments{
\item{index}{A column that identifies the time variable. Usually a date column.}

\item{keys}{A set of colums that uniquely identify each time series.}

\item{static}{A set of colums that will be treated as static predictors by the
model. Ie, those are variables that don't vary in time.}

\item{known}{A set of columns that are known for every possible input \code{index}.
Examples are 'day of the week' or 'is_holiday' flags that are always known even
for future timesteps.}

\item{unknown}{A set of columns that are considered to be unknown variables.
By default, all columns that don't fit in any of the previous parameters are
considered unknown, so you don't need to specify it manually.}
}
\description{
Create a specification of covariates types
}
