% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials.R
\name{dataprep_parameters}
\alias{dataprep_parameters}
\alias{lookback}
\alias{horizon}
\alias{hidden_state_size}
\title{Parameters related to data preparation in TFT}
\usage{
lookback(range = c(3L, 365L), trans = NULL)

horizon(range = c(1L, 365L), trans = NULL)

hidden_state_size(range = c(2L, 128L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Parameters related to data preparation in TFT
}
\section{Functions}{
\itemize{
\item \code{lookback}: lookback Lookback from the history

\item \code{horizon}: horizon Number of steps in the multi-horizon forecast

\item \code{hidden_state_size}: hidden_state_size Size of network.
}}

\examples{
lookback()
}
