% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{fit}
\alias{fit}
\title{Fit a model
See \code{\link[generics:fit]{generics::fit()}} for more information.}
\description{
Fit a model
See \code{\link[generics:fit]{generics::fit()}} for more information.
}
\keyword{internal}
