% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit.tft_module}
\alias{fit.tft_module}
\title{Fit the Temporal Fusion Transformer module}
\usage{
\method{fit}{tft_module}(object, ...)
}
\arguments{
\item{object}{a TFT module created with \code{\link[=temporal_fusion_transformer]{temporal_fusion_transformer()}}.}

\item{...}{Arguments passed to \code{\link[luz:fit.luz_module_generator]{luz::fit.luz_module_generator()}}.}
}
\description{
Fit the Temporal Fusion Transformer module
}
