% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{forecast}
\alias{forecast}
\title{Creates forecasts
See \code{\link[generics:forecast]{generics::forecast()}} for more information.}
\description{
Creates forecasts
See \code{\link[generics:forecast]{generics::forecast()}} for more information.
}
\keyword{internal}
