% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy-tsibble.R
\name{known_input_rules}
\alias{known_input_rules}
\title{Provide rules to define variables that are known for future timesteps.}
\usage{
known_input_rules(.data, ...)
}
\arguments{
\item{.data}{a tsibble}

\item{...}{functions that are used to define new variables based only on
the index and the keys of \code{.data} tsibble.}
}
\description{
Provide rules to define variables that are known for future timesteps.
}
\examples{

library(dplyr)
library(tsibble)
weather <- nycflights13::weather \%>\%
  select(origin, time_hour, temp, humid, precip)
weather_tsbl <- as_tsibble(weather, key = origin, index = time_hour)
weather_tsbl

weather_tsbl \%>\%
  known_input_rules(
    hour = function(x, ...) lubriate::hour(x)
  ) \%>\%
  new_data()

}
