% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tft}
\alias{predict.tft}
\title{Create predictions for TFT models}
\usage{
\method{predict}{tft}(object, new_data, type = "numeric", mode = "horizon", step = NULL, ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{new_data}{A \code{\link[=data.frame]{data.frame()}} containing a dataset to generate predictions
for. In general it's used to pass static and known information to generate
forecasts.}

\item{type}{Currently only \code{'numeric'} is accepted but this might change in
the future if we end up supporting classification.}

\item{mode}{Prediction mode. If \code{'horizon'} predict will generate a single
multi-horizon prediction. This is mostly sueful when creating forecasts for
the time frame right after the model has been trained. If \code{'full'} then
predictions are created for every possible time step (with a possible \code{step})
argument.
When \code{mode='horizon'} (default) only the prediction columns are returned.
When \code{mode='full'} all columns from \code{new_data} are returned as the result
might have more lines than in \code{new_data} because it's potentially possible
to generate different predictions for the same time-step.}

\item{step}{Step for predictions when using \code{mode='full'}.}

\item{...}{additional arguments affecting the predictions produced.}
}
\description{
Create predictions for TFT models
}
