% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tft_result}
\alias{predict.tft_result}
\title{Predict for TFT}
\usage{
\method{predict}{tft_result}(object, new_data = NULL, ..., past_data = NULL)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{new_data}{A \code{\link[=data.frame]{data.frame()}} containing a dataset to generate predictions
for. In general it's used to pass static and known information to generate
forecasts.}

\item{...}{other arguments passed to the \code{\link[=predict.luz_module_fitted()]{predict.luz_module_fitted()()}}.}

\item{past_data}{A \code{\link[=data.frame]{data.frame()}} with past information for creating the
predictions. It should include at least \code{lookback} values - but can be more.
It's concatenated with \code{new_data} before passing forward. If \code{NULL}, the
data used to train the model is used.}
}
\description{
Predict for TFT
}
