% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{prep}
\alias{prep}
\title{Prepare a specification
See \code{\link[recipes:prep]{recipes::prep()}} for more information.}
\description{
Prepare a specification
See \code{\link[recipes:prep]{recipes::prep()}} for more information.
}
\keyword{internal}
