% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series-dataset.R
\name{spec_time_splits}
\alias{spec_time_splits}
\title{Time splits setting}
\usage{
spec_time_splits(spec, lookback, horizon)
}
\arguments{
\item{lookback}{Number of timesteps that are used as historic data for
prediction.}

\item{horizon}{Number of timesteps ahead that will be predicted by the
model.}
}
\description{
Time splits setting
}
