% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-normalize.R
\name{step_group_normalize}
\alias{step_group_normalize}
\title{Group normalization}
\usage{
step_group_normalize(
  recipe,
  ...,
  groups,
  stats = NULL,
  role = NA,
  trained = FALSE,
  skip = FALSE,
  id = recipes::rand_id("group_normalize")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[recipes:selections]{selections()}} for more details.}

\item{groups}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to group by
before computing normalization statistics.}

\item{stats}{Is modified by \code{prep}. A data frame containing one row per distinct
group, containing the normalization statistics.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{bake()}}? While all operations are baked
when \code{\link[recipes:prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
\code{\link[recipes:recipe]{recipes::recipe()}} step for normalizing data per group.
Most of the times we want to normalize each time series independently as
they might have different scales.
}
