% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-roles.R
\name{step_include_roles}
\alias{step_include_roles}
\title{Include roles}
\usage{
step_include_roles(
  recipe,
  roles = NULL,
  trained = FALSE,
  skip = FALSE,
  role = NA,
  id = recipes::rand_id("include_roles")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{roles}{a data.frame with role information for each term created by the
recipe.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{bake()}}? While all operations are baked
when \code{\link[recipes:prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{role}{Not used by this step since no new variables are
created.}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
A \code{\link[recipes:recipe]{recipes::recipe()}} step that adds role information and the recipe itself
as an attribute of the \code{bake}d or \code{juice}d output.
}
\details{
This step must be used as the last step in a recipe.
}
