% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{temporal_fusion_transformer}
\alias{temporal_fusion_transformer}
\alias{tft_config}
\title{Temporal Fusion transformer}
\usage{
temporal_fusion_transformer(spec, ...)

tft_config(
  hidden_state_size = 16,
  num_attention_heads = 4,
  num_lstm_layers = 2,
  dropout = 0.1,
  optimizer = "adam",
  learn_rate = 0.01,
  quantiles = c(0.1, 0.5, 0.9)
)
}
\arguments{
\item{spec}{A spec created with \code{\link[=tft_dataset_spec]{tft_dataset_spec()}}.
This is required because the model depends on some information that is
created/defined in the dataset.}

\item{...}{Additional parameters passed to \code{\link[=tft_config]{tft_config()}}.}

\item{hidden_state_size}{Hidden size of network which is its main hyperparameter
and can range from 8 to 512. It's also known as \code{d_model} across the paper.}

\item{num_attention_heads}{Number of attention heads in the Multi-head attention layer.
The paper refer to it as \code{m_H}. \code{4} is a good default.}

\item{num_lstm_layers}{Number of LSTM layers used in the Locality Enhancement
Layer. Usually 2 is good enough.}

\item{dropout}{Dropout rate used in many places in the architecture.}

\item{optimizer}{Optimizer used for training. Can be a string with 'adam', 'sgd',
or 'adagrad'. Can also be a \code{\link[torch:optimizer]{torch::optimizer()}}.}

\item{learn_rate}{Leaning rate used by the optimizer.}

\item{quantiles}{A numeric vector with 3 quantiles for the quantile loss.
The first is treated as lower bound of the interval, the second as the
point prediction and the thir as the upper bound.}
}
\value{
A \code{luz_module} that has been setup and is ready to be \code{fitted}.
}
\description{
Temporal Fusion transformer

Configuration for the tft model
}
\section{Functions}{
\itemize{
\item \code{temporal_fusion_transformer()}: Create the tft module

\item \code{tft_config()}: Configuration for the Temporal Fusion Transformer

}}
\seealso{
\code{\link[=fit.luz_module_generator]{fit.luz_module_generator()}} for fit arguments. See
\code{\link[=predict.tft_result]{predict.tft_result()}} for information on how to generate predictions and
\code{\link[=forecast.tft_result]{forecast.tft_result()}} for forecasting.
}
