% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tft}
\alias{predict.tft}
\title{Create predictions for TFT models}
\usage{
\method{predict}{tft}(object, new_data, type = "numeric", ..., past_data = NULL)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{new_data}{A \code{\link[=data.frame]{data.frame()}} containing a dataset to generate predictions
for. In general it's used to pass static and known information to generate
forecasts.}

\item{type}{Currently only \code{'numeric'} is accepted but this might change in
the future if we end up supporting classification.}

\item{...}{additional arguments affecting the predictions produced.}

\item{step}{Step for predictions when using \code{mode='full'}.}
}
\description{
Create predictions for TFT models
}
