% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{rolling_predict}
\alias{rolling_predict}
\title{Defines rolling slices}
\usage{
rolling_predict(object, past_data, new_data, step = NULL)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{past_data}{A \code{\link[=data.frame]{data.frame()}} with past information for creating the
predictions. It should include at least \code{lookback} values - but can be more.
It's concatenated with \code{new_data} before passing forward. If \code{NULL}, the
data used to train the model is used.}

\item{new_data}{A \code{\link[=data.frame]{data.frame()}} containing a dataset to generate predictions
for. In general it's used to pass static and known information to generate
forecasts.}

\item{step}{Default is the step to be the same as the horizon o the model,
that way we have one prediction per slice.}
}
\description{
Sometimes your validation or testing data has more values than the \code{horizon}
of your model but you still want to create predictions for each time step on
them.
}
\details{
This function will combine your \code{past_data} (that can also include your training data)
and create slices so you create predictions for each value in \code{new_data}.
}
